pyproject.toml
[project]
name = "ai-marketing-agents"
version = "0.1.0"
description = "AI-powered social media marketing platform"
requires-python = ">=3.11"
dependencies = [
    # --- Orchestration ---
    "langgraph>=0.2.0",
    "langgraph-checkpoint-mysql>=1.0.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-google-genai>=2.0.0",

    # --- Telegram Bot ---
    "python-telegram-bot[ext]>=21.0",

    # --- Database ---
    "sqlalchemy[asyncio]>=2.0.0",
    "aiomysql>=0.2.0",
    "pymysql>=1.1.0",
    "alembic>=1.13.0",

    # --- AI / LLM ---
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "google-generativeai>=0.8.0",
    "instructor>=1.4.0",

    # --- Image Processing ---
    "Pillow>=10.0.0",
    "opencv-python-headless>=4.9.0",
    "rembg>=2.0.0",

    # --- Video Processing ---
    "ffmpeg-python>=0.2.0",

    # --- Voiceover ---
    "edge-tts>=6.1.0",

    # --- Speech-to-Text ---
    "openai-whisper>=20231117",

    # --- Vector DB ---
    "chromadb>=0.5.0",

    # --- Hashtag & Trends ---
    "google-search-results>=2.4.0",

    # --- Document Parsing (Resumes) ---
    "PyMuPDF>=1.24.0",
    "python-docx>=1.1.0",

    # --- Calendar ---
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",

    # --- Social Media APIs ---
    "facebook-sdk>=3.1.0",
    "requests>=2.31.0",
    "requests-oauthlib>=1.3.0",
    "httpx>=0.27.0",

    # --- Audio Manipulation ---
    "pydub>=0.25.0",

    # --- Utilities ---
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "aiofiles>=23.2.0",
    "structlog>=23.2.0",
    "redis>=5.0.0",
]

[project.scripts]
marketing-bot = "bot.main:main"
